Remove unpromoted alpha features, organize promoted alpha features
