--- internal/ansible/ansible_suite_test.go
@@ -1,13 +1,0 @@
-package ansible_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestAnsible(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Ansible Suite")
-}
