--- internal/cmd/productctl/cmd/certifycontainers/certifycontainers_suite_test.go
@@ -1,13 +1,0 @@
-package certifycontainers_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestCertifycontainers(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Certifycontainers Suite")
-}
