--- internal/cmd/productctl/cmd/certifyoperators/certifyoperators_suite_test.go
@@ -1,13 +1,0 @@
-package certifyoperators_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestCertifyoperators(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Certifyoperators Suite")
-}
