--- internal/cmd/productctl/cmd/certtargets/certtargets_suite_test.go
@@ -1,13 +1,0 @@
-package certtargets_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestCerttargets(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Certtargets Suite")
-}
